+++ docs/en/changes/changes.md
@@ -1,18 +1,23 @@
 ## 9.1.0
 
 #### Project
+
 * Upgrade zipkin to 2.23.16.
 
 #### OAP Server
+
 * Add component definition(ID=127) for `Apache ShenYu (incubating)`.
-* Fix Zipkin receiver: Decode spans error, missing `Layer` for V9 and wrong time bucket for generate Service and Endpoint.
+* Fix Zipkin receiver: Decode spans error, missing `Layer` for V9 and wrong time bucket for generate Service and
+  Endpoint.
+* [Refactor] Move ElasticSearch and BanyanDB specific configurations out of column.
+* Support BanyanDB global index for entities. Log and Segment record entities declare this new feature.
 
 #### UI
+
 * General service instance: move `Thread Pool` from JVM to Overview, fix `JVM GC Count` calculation.
 * Add Apache ShenYu (incubating) component LOGO.
 
 #### Documentation
-
 
 All issues and pull requests are [here](https://github.com/apache/skywalking/milestone/128?closed=1)